+++ changelog.md
@@ -1,4 +1,9 @@
 # Changelog
 
+- 2023-07-31
+  - Added touch handler.
+- 2023-07-26
+  - Added editor.
+  - Added mouse dragging of boxes.
 - 2023-07-25
   - Init.